# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Soroticket Testnet Cloud API",
    "version": "0.2.1-rc.1",
    "description": "Tranche 1 baseline. Testnet only; TEST and METERED never move mainnet value. The console uses the same-origin /api prefix. API key auth or an HttpOnly session is accepted on /v1. Browser session mutations require same origin. Publish no PII in campaign names/codes. Off-chain receipts attest events, not physical truth."
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local API"
    },
    {
      "url": "/api",
      "description": "Hosted console, same origin"
    }
  ],
  "security": [
    {
      "ApiKey": []
    },
    {
      "Session": []
    }
  ],
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "post_auth_signup",
        "summary": "Create account and session",
        "description": "Create account and session",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "summary": "Sign in",
        "description": "Sign in",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "summary": "Sign out",
        "description": "Sign out",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Session": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "get_auth_me",
        "summary": "Current account and organization",
        "description": "Current account and organization",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "Session": []
          }
        ]
      }
    },
    "/orgs": {
      "post": {
        "operationId": "post_orgs",
        "summary": "Create organization and start testnet account funding",
        "description": "Create organization and start testnet account funding",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 96
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "security": [
          {
            "Session": []
          }
        ]
      }
    },
    "/v1/overview": {
      "get": {
        "operationId": "get_v1_overview",
        "summary": "Overview including account_funded",
        "description": "Overview including account_funded",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/keys": {
      "get": {
        "operationId": "get_v1_keys",
        "summary": "List API keys without their secrets",
        "description": "List API keys without their secrets",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_v1_keys",
        "summary": "Create API key returned once",
        "description": "Create API key returned once",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Reuse with the exact same body for retries. Conflicting reuse or in-progress operation returns 409. Replay retained 24 hours."
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/keys/{id}/revoke": {
      "post": {
        "operationId": "post_v1_keys_id_revoke",
        "summary": "Revoke owned API key",
        "description": "Revoke owned API key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Reuse with the exact same body for retries. Conflicting reuse or in-progress operation returns 409. Replay retained 24 hours."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/campaigns": {
      "get": {
        "operationId": "get_v1_campaigns",
        "summary": "List tenant campaigns",
        "description": "List tenant campaigns",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaigns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Campaign"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_v1_campaigns",
        "summary": "Create Burn or Tally campaign",
        "description": "ticket/voucher require positive supply. coupon/creator/gift require shared.code; creator also requires attributed_to. Names and codes are public. Wait for account_funded before invoking.",
        "parameters": [
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Reuse with the exact same body for retries. Conflicting reuse or in-progress operation returns 409. Replay retained 24 hours."
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignInput"
              }
            }
          }
        }
      }
    },
    "/v1/campaigns/{id}": {
      "get": {
        "operationId": "get_v1_campaigns_id",
        "summary": "Campaign and issued codes",
        "description": "Campaign and issued codes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/campaigns/{id}/archive": {
      "post": {
        "operationId": "post_v1_campaigns_id_archive",
        "summary": "Archive campaign in Cloud",
        "description": "Archive campaign in Cloud",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Reuse with the exact same body for retries. Conflicting reuse or in-progress operation returns 409. Replay retained 24 hours."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/campaigns/{id}/codes": {
      "post": {
        "operationId": "post_v1_campaigns_id_codes",
        "summary": "Issue up to 1000 unique codes in batches of 100",
        "description": "Issue up to 1000 unique codes in batches of 100",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-Env",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "live"
              ],
              "default": "test"
            },
            "description": "For session auth. live is the internal name of the metered testnet preview; API keys select their own environment."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Reuse with the exact same body for retries. Conflicting reuse or in-progress operation returns 409. Replay retained 24 hours."
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 authentication; 403 authorization/CSRF; 404 absent; 409 conflict; 429 rate limit; 500 error; 503 funding/not-ready. Do not blindly retry an ambiguous chain write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "207": {
            "description": "Partial issuance. Inspect issued[] and error; do not repeat already-issued codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "sk_test_\u2026 key; full key is returned only on creation."
      },
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sd_session"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "Credentials": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 15,
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "IssueInput": {
        "type": "object",
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Public code; maximum 64 UTF-8 bytes. Never include PII."
            },
            "maxItems": 1000
          },
          "generate": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer",
                "minimum": 1,
                "maximum": 1000
              },
              "prefix": {
                "type": "string"
              }
            },
            "required": [
              "count"
            ]
          }
        }
      },
...
    }
  }
}